fix: Force import "kolmafia" for tsc --build
tsc --build relies on d.ts files generated between different projects.
Unfortunately, when tsc generates type declarations (DTS) for
parse-displaycollection.ts, it entirely removes statements like this:

    import {X} from "kolmafia";

However, this import has an important side effect of introducing
KoLmafia's globals (Item, Element, etc.), and must be present in the
emitted DTS for it to be valid.

To emit a valid DTS, we must add a duplicate import statement:

    import "kolmafia";
    import {X} from "kolmafia";

This ensures that our type definitions are still valid.

One minor problem is that "kolmafia" is now imported twice in the
emitted JS file. This is not a problem, however, since it seems that
Rollup removes it anyway during bundling.