fix: Use CSS Flexbox instead of Grid
In Google Chrome, CSS grids with >1000 rows do not render correctly.
This is an intentional limit. For more information, see:
  https://bugs.chromium.org/p/chromium/issues/detail?id=688640

This becomes a problem because some Display Cases contain thousands of
items in a single shelf.

As a workaround, use CSS Flexbox instead of Grid. Use media queries to
maintain a responsive layout that auto-collapses when the viewport
becomes narrow.