fix: Don't select DC message containing <font> tag
Display Case messages may contain a <font> tag if the user puts a fully
qualified URL (e.g. http://www.example.com/). This caused the previous
XPath selector to incorrectly parse the display case message field as a
shelf.

Fix this by selecting a <table> that contains a <span> with an `id`
attribute.